fixed a cast
git-svn-id: https://pimpmobile.svn.sourceforge.net/svnroot/pimpmobile/trunk@40 3d5ecaf0-f903-0410-b953-c2c1a4d75763

+++ converter_mod.cpp
@@ -14,7 +14,7 @@ int return_nearest_note(int p)
 	double log2p = log(p)/log(2);
 	double note = (log2( 428 << 5) - log2p) * 12.0 + 1;
 	
-	int note_int = floor( note + .5 );
+	int note_int = int(floor( note + .5 ));
 
 	if (note_int <= 0 || note_int > 120) return 0; // this is not a note
 	return note_int;
